Trait vested_rewards::vesting_currencies::VestingSchedule

pub trait VestingSchedule<BlockNumber, Balance, AssetId: Copy> {
    // Required methods
    fn end(&self) -> Option<BlockNumber>;
    fn total_amount(&self) -> Option<Balance>;
    fn locked_amount(&self, time: BlockNumber) -> Option<Balance>;
    fn ensure_valid_vesting_schedule<T: Config>(
        &self
    ) -> Result<Balance, DispatchError>;
    fn asset_id(&self) -> AssetId;
    fn next_claim_block<T: Config>(
        &self,
        time: BlockNumber
    ) -> Result<Option<BlockNumber>, DispatchError>;
    fn claims_count(&self) -> u32;
}

Required Methods

fn end(&self) -> Option<BlockNumber>

Returns the end of all periods, None if calculation overflows.

fn total_amount(&self) -> Option<Balance>

Returns all locked amount, None if calculation overflows.

fn locked_amount(&self, time: BlockNumber) -> Option<Balance>

Returns locked amount for a given time.

fn ensure_valid_vesting_schedule<T: Config>( &self ) -> Result<Balance, DispatchError>

fn asset_id(&self) -> AssetId

Returns asset id, need to get from enum

fn next_claim_block<T: Config>( &self, time: BlockNumber ) -> Result<Option<BlockNumber>, DispatchError>

Returns next block for a given time, where asset may be unlocked and claimed

fn claims_count(&self) -> u32

Count of claims per Vesting